+++ client/app2/components/testcase.htm
@@ -12,7 +12,7 @@
           title="Save your Test Case changes">
         <i class="fa fa-floppy-o" aria-hidden="true"></i>
       </a> &nbsp;
-      <a href="" class="icon-download" ng-click="vm.downloadTestCase()"
+      <a href="" class="icon-download" ng-click="vm.downloadTestcase()"
           title="Download Selenium script for this Test Case">
         <i class="fa fa-download" aria-hidden="true"></i>
       </a> &nbsp;
@@ -182,7 +182,7 @@
   <div class="col-md-6">
   </div>
 	<div class="col-md-6 text-right">
-		<button class="btn btn-success" ng-click="vm.downloadTestCase()">
+		<button class="btn btn-success" ng-click="vm.downloadTestcase()">
 			<i class="fa fa-download" aria-hidden="true"></i>
 			Download Test Case
 		</button>
